changed number type to decimal, updated tests

+++ Fluid/DefaultFluidParser.cs
@@ -1030,7 +1030,7 @@ namespace Fluid
 
                 case "number":
                     var decimalSeparator = System.Globalization.CultureInfo.CurrentCulture.NumberFormat.NumberDecimalSeparator;
-                    return new LiteralExpression(NumberValue.Create(Convert.ToDouble(node.Token.Value), !node.Token.Text.Contains(decimalSeparator)));
+                    return new LiteralExpression(NumberValue.Create(Convert.ToDecimal(node.Token.Value), !node.Token.Text.Contains(decimalSeparator)));
 
                 case "boolean":
                     if (!bool.TryParse(node.ChildNodes[0].Token.Text, out var boolean))
